Add `--no-lint` flag to `bump-version.bash`
On my system, I run build related tasks in Docker. While this works for
linting, I am unable to bump the version from within that environment
due to (a) lack of EDITOR and (b) no access to my git credentials. This
flag just skips the linting step if it is provided.

I think it makes sense to introduce this flag, at least in my usecase.
If used responsibly (e.g. linting before running this) then things
should be fine.