Add a little bit of guidance on GH Action setup
The GH Action workflow requires 3 secrets configured so that the
application can be built and signed. In the guide, for now, I have just
named the secrets that I have defined, since I forget how exactly I
generated them.

That being said, they are pretty standard secrets to have when
developing Android applications. Hopefully this amount of information is
helpful to someone.